update: more privacy settings options available, pending integration
