+++ app/src/main/java/com/example/sw0b_001/SettingsActivities/SecurityPrivacyFragment.java
@@ -1,0 +1,15 @@
+package com.example.sw0b_001.SettingsActivities;
+
+import android.os.Bundle;
+
+import androidx.preference.PreferenceFragmentCompat;
+
+import com.example.sw0b_001.R;
+
+public class SecurityPrivacyFragment extends PreferenceFragmentCompat {
+
+    @Override
+    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
+        setPreferencesFromResource(R.xml.security_privacy_preferences, rootKey);
+    }
+}
