--- app/src/main/java/com/example/sw0b_001/SettingsFragment.java
@@ -1,28 +1,0 @@
-package com.example.sw0b_001;
-
-import android.content.Intent;
-import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-import androidx.preference.Preference;
-import androidx.preference.PreferenceFragmentCompat;
-
-public class SettingsFragment extends PreferenceFragmentCompat {
-
-    @Override
-    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
-        setPreferencesFromResource(R.xml.settings_preferences, rootKey);
-
-        Preference languagePreference = findPreference("language_settings");
-        languagePreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-            @Override
-            public boolean onPreferenceClick(@NonNull Preference preference) {
-                Intent languageIntent = new Intent(getContext(), LanguageSettingsActivity.class);
-                startActivity(languageIntent);
-                return false;
-            }
-        });
-    }
-
-
-}
