+++ app/src/main/res/layout/activity_settings.xml
@@ -24,7 +24,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container_view"
-        android:name="com.example.sw0b_001.SettingsFragment"
+        android:name="com.example.sw0b_001.SettingsActivities.SettingsFragment"
         android:layout_width="wrap_content"
         android:layout_height="0dp"
         android:layout_marginTop="8dp"
