+++ app/src/main/res/values/strings.xml
@@ -190,18 +190,10 @@
     <string name="settings_security_and_privacy_enable_pin">
         Enable PIN for decryption
     </string>
-    <!-- Preference Titles -->
-    <string name="messages_header">Messages</string>
-    <string name="sync_header">Sync</string>
-
-    <!-- Messages Preferences -->
-    <string name="signature_title">Your signature</string>
-    <string name="reply_title">Default reply action</string>
-
-    <!-- Sync Preferences -->
-    <string name="sync_title">Sync email periodically</string>
-    <string name="attachment_title">Download incoming attachments</string>
-    <string name="attachment_summary_on">Automatically download attachments for incoming emails
+    <string name="settings_lock_screen_preference_category">
+        Phone Lockscreen options
     </string>
-    <string name="attachment_summary_off">Only download attachments when manually requested</string>
+    <string name="settings_pin_preference_category">
+        Custom PIN options
+    </string>
 </resources>
